bump packages

+++ test/src/csv.ts
@@ -1,9 +1,9 @@
 import {Column, Table, Types} from '@subsquid/file-store-csv'
 
 export const Transfers = new Table('transfers.csv', {
-    blockNumber: Column(Types.Integer()),
+    blockNumber: Column(Types.Numeric()),
     timestamp: Column(Types.DateTime()),
     from: Column(Types.String()),
     to: Column(Types.String()),
-    amount: Column(Types.Integer()),
+    amount: Column(Types.Numeric()),
 })
